Remove integration tests for Segformer and Enable weights from CIv2
