+++ .github/workflows/perf-device-models-impl.yaml
@@ -68,7 +68,7 @@ jobs:
             pytest models/demos/convnet_mnist/tests/ -m models_device_performance_bare_metal
             pytest models/demos/mnist/tests -m models_device_performance_bare_metal
             WH_ARCH_YAML=$MAGIC_ENV pytest models/demos/falcon7b_common/tests -m models_device_performance_bare_metal
-            WH_ARCH_YAML=$MAGIC_ENV pytest models/demos/segformer/tests -m models_device_performance_bare_metal
+            WH_ARCH_YAML=$MAGIC_ENV pytest models/demos/segformer/tests/perf/ -m models_device_performance_bare_metal
         if: ${{ !cancelled() && !matrix.test-info.civ2-compatible }}
         timeout-minutes: ${{ matrix.test-info.timeout }}
