+++ .github/workflows/perf-models-impl.yaml
@@ -60,7 +60,7 @@ jobs:
               # This is bad. This should be referencing .../tests directory
               - models/demos/wormhole/distilbert/tests/test_perf_distilbert.py
               # This is bad. This should be referencing .../tests directory
-              - models/demos/segformer/tests/test_e2e_performant.py
+              - models/demos/segformer/tests/perf/
               - models/demos/metal_BERT_large_11/tests/
               - models/demos/yolov8s_world/tests/perf/
               - models/demos/yolov8s/tests/perf/
