+++ models/demos/segformer/configs/segformer_encoder_config.json
@@ -1,0 +1,118 @@
+{
+    "return_dict": true,
+    "output_hidden_states": false,
+    "torchscript": false,
+    "torch_dtype": null,
+    "use_bfloat16": false,
+    "tf_legacy_loss": false,
+    "pruned_heads": {},
+    "tie_word_embeddings": true,
+    "chunk_size_feed_forward": 0,
+    "is_encoder_decoder": false,
+    "is_decoder": false,
+    "cross_attention_hidden_size": null,
+    "add_cross_attention": false,
+    "tie_encoder_decoder": false,
+    "max_length": 20,
+    "min_length": 0,
+    "do_sample": false,
+    "early_stopping": false,
+    "num_beams": 1,
+    "num_beam_groups": 1,
+    "diversity_penalty": 0.0,
+    "temperature": 1.0,
+    "top_k": 50,
+    "top_p": 1.0,
+    "typical_p": 1.0,
+    "repetition_penalty": 1.0,
+    "length_penalty": 1.0,
+    "no_repeat_ngram_size": 0,
+    "encoder_no_repeat_ngram_size": 0,
+    "bad_words_ids": null,
+    "num_return_sequences": 1,
+    "output_scores": false,
+    "return_dict_in_generate": false,
+    "forced_bos_token_id": null,
+    "forced_eos_token_id": null,
+    "remove_invalid_values": false,
+    "exponential_decay_length_penalty": null,
+    "suppress_tokens": null,
+    "begin_suppress_tokens": null,
+    "architectures": null,
+    "finetuning_task": null,
+    "id2label": {
+        "0": "LABEL_0",
+        "1": "LABEL_1"
+    },
+    "label2id": {
+        "LABEL_0": 0,
+        "LABEL_1": 1
+    },
+    "tokenizer_class": null,
+    "prefix": null,
+    "bos_token_id": null,
+    "pad_token_id": null,
+    "eos_token_id": null,
+    "sep_token_id": null,
+    "decoder_start_token_id": null,
+    "task_specific_params": null,
+    "problem_type": null,
+    "_name_or_path": "",
+    "transformers_version": "4.53.0",
+    "num_channels": 3,
+    "num_encoder_blocks": 4,
+    "depths": [
+        2,
+        2,
+        2,
+        2
+    ],
+    "sr_ratios": [
+        8,
+        4,
+        2,
+        1
+    ],
+    "hidden_sizes": [
+        32,
+        64,
+        160,
+        256
+    ],
+    "patch_sizes": [
+        7,
+        3,
+        3,
+        3
+    ],
+    "strides": [
+        4,
+        2,
+        2,
+        2
+    ],
+    "mlp_ratios": [
+        4,
+        4,
+        4,
+        4
+    ],
+    "num_attention_heads": [
+        1,
+        2,
+        5,
+        8
+    ],
+    "hidden_act": "gelu",
+    "hidden_dropout_prob": 0.0,
+    "attention_probs_dropout_prob": 0.0,
+    "classifier_dropout_prob": 0.1,
+    "initializer_range": 0.02,
+    "drop_path_rate": 0.1,
+    "layer_norm_eps": 1e-06,
+    "decoder_hidden_size": 256,
+    "reshape_last_stage": true,
+    "semantic_loss_ignore_index": 255,
+    "model_type": "segformer",
+    "output_attentions": false
+}
