+++ models/demos/segformer/tests/pcc/test_segformer_attention.py
@@ -2,25 +2,22 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import torch
-import ttnn
-from ttnn.model_preprocessing import (
-    preprocess_model_parameters,
-)
-from transformers import SegformerModel
 import pytest
-from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.utility_functions import skip_for_grayskull
-from models.demos.segformer.tt.ttnn_segformer_attention import (
-    TtSegformerAttention,
-)
+import torch
+from ttnn.model_preprocessing import preprocess_model_parameters
+
+import ttnn
+from models.demos.segformer.common import load_config, load_torch_model
 from models.demos.segformer.reference.segformer_attention import SegformerAttention
-from tests.ttnn.integration_tests.segformer.test_segformer_efficient_selfattention import (
+from models.demos.segformer.tests.pcc.test_segformer_efficient_selfattention import (
     create_custom_preprocessor as create_customer_preprocessor_selfattention,
 )
-from tests.ttnn.integration_tests.segformer.test_segformer_selfoutput import (
+from models.demos.segformer.tests.pcc.test_segformer_selfoutput import (
     create_custom_preprocessor as create_customer_preprocessor_selfoutput,
 )
+from models.demos.segformer.tt.ttnn_segformer_attention import TtSegformerAttention
+from models.utility_functions import skip_for_grayskull
+from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
 def create_custom_preprocessor(device):
@@ -66,8 +63,7 @@ def test_segformer_attention(
     width,
     block_i,
     attention_i,
-    reset_seeds,
-    is_ci_env,
+    model_location_generator,
 ):
     torch_input_tensor = torch.randn(batch_size, 1, seq_len, hidden_size)
     ttnn_input_tensor = ttnn.from_torch(
@@ -77,10 +73,10 @@ def test_segformer_attention(
         device=device,
         layout=ttnn.TILE_LAYOUT,
     )
-    torch_model = SegformerModel.from_pretrained("nvidia/segformer-b0-finetuned-ade-512-512")
-    config = torch_model.config
-    torch_model = torch_model.encoder.block[block_i][attention_i].attention
 
+    target_prefix = f"segformer.encoder.block.{block_i}.{attention_i}.attention."
+
+    config = load_config("configs/segformer_semantic_config.json")
     reference_model = SegformerAttention(
         config,
         hidden_size=hidden_size,
@@ -88,9 +84,9 @@ def test_segformer_attention(
         sequence_reduction_ratio=sequence_reduction_ratio,
     )
 
-    sd = torch_model.state_dict()
-    reference_model.load_state_dict(sd)
-    reference_model.eval()
+    reference_model = load_torch_model(
+        reference_model, target_prefix, module="semantic_sub", model_location_generator=model_location_generator
+    )
 
     torch_input_tensor = torch.reshape(torch_input_tensor, (batch_size, seq_len, hidden_size))
     output = reference_model(torch_input_tensor, height, width)
