+++ models/demos/segformer/tests/pcc/test_segformer_dwconv.py
@@ -2,16 +2,16 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
+import pytest
 import torch
-import ttnn
 from ttnn.model_preprocessing import preprocess_model_parameters
 
-from tests.ttnn.utils_for_testing import assert_with_pcc
-from transformers import SegformerModel
+import ttnn
+from models.demos.segformer.common import load_torch_model
 from models.demos.segformer.reference.segformer_dwconv import SegformerDWConv
-import pytest
 from models.demos.segformer.tt.ttnn_segformer_dwconv import TtSegformerDWConv
 from models.utility_functions import skip_for_grayskull
+from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
 def create_custom_preprocessor(device):
@@ -44,7 +44,9 @@ def create_custom_preprocessor(device):
     ],
 )
 @pytest.mark.parametrize("device_params", [{"l1_small_size": 16384}], indirect=True)
-def test_segformer_dw_conv(device, batch_size, seq_len, dim, height, width, block_i, dwconv_i, reset_seeds, is_ci_env):
+def test_segformer_dw_conv(
+    device, batch_size, seq_len, dim, height, width, block_i, dwconv_i, model_location_generator
+):
     torch_input_tensor = torch.randn(batch_size, seq_len, dim)
     ttnn_input_tensor = ttnn.from_torch(
         torch_input_tensor,
@@ -53,13 +55,12 @@ def test_segformer_dw_conv(device, batch_size, seq_len, dim, height, width, bloc
         device=device,
         layout=ttnn.TILE_LAYOUT,
     )
-    torch_model = SegformerModel.from_pretrained("nvidia/segformer-b0-finetuned-ade-512-512")
-    torch_model = torch_model.encoder.block[block_i][dwconv_i].mlp.dwconv
 
     reference_model = SegformerDWConv(dim=dim)
-    sd = torch_model.state_dict()
-    reference_model.load_state_dict(sd)
-    reference_model.eval()
+    target_prefix = f"encoder.block.{block_i}.{dwconv_i}.mlp.dwconv."
+    reference_model = load_torch_model(
+        reference_model, target_prefix, module="semantic_sub", model_location_generator=model_location_generator
+    )
 
     torch_output = reference_model(torch_input_tensor, height, width)
