+++ models/demos/segformer/tests/pcc/test_segformer_efficient_selfattention.py
@@ -2,24 +2,21 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
+import pytest
 import torch
-import ttnn
 from ttnn.model_preprocessing import (
-    preprocess_model_parameters,
     preprocess_layernorm_parameter,
     preprocess_linear_bias,
     preprocess_linear_weight,
+    preprocess_model_parameters,
 )
-from transformers import SegformerModel
-import pytest
-from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.demos.segformer.tt.ttnn_segformer_efficient_selfattention import (
-    TtSegformerEfficientSelfAttention,
-)
+
+import ttnn
+from models.demos.segformer.common import load_config, load_torch_model
+from models.demos.segformer.reference.segformer_efficient_selfattention import SegformerEfficientSelfAttention
+from models.demos.segformer.tt.ttnn_segformer_efficient_selfattention import TtSegformerEfficientSelfAttention
 from models.utility_functions import skip_for_grayskull
-from models.demos.segformer.reference.segformer_efficient_selfattention import (
-    SegformerEfficientSelfAttention,
-)
+from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
 def create_custom_preprocessor(device):
@@ -85,8 +82,7 @@ def test_segformer_efficient_selfattention(
     sequence_reduction_ratio,
     block_i,
     efficient_self_attention_i,
-    reset_seeds,
-    is_ci_env,
+    model_location_generator,
 ):
     torch_input_tensor = torch.randn(batch_size, 1, seq_len, hidden_size)
     ttnn_input_tensor = ttnn.from_torch(
@@ -96,20 +92,18 @@ def test_segformer_efficient_selfattention(
         memory_config=ttnn.L1_MEMORY_CONFIG,
         layout=ttnn.TILE_LAYOUT,
     )
-    torch_model = SegformerModel.from_pretrained("nvidia/segformer-b0-finetuned-ade-512-512")
-    config = torch_model.config
-    torch_model = torch_model.encoder.block[block_i][efficient_self_attention_i].attention.self
 
+    config = load_config("configs/segformer_semantic_config.json")
     reference_model = SegformerEfficientSelfAttention(
         config=config,
         hidden_size=hidden_size,
         num_attention_heads=num_attention_heads,
         sequence_reduction_ratio=sequence_reduction_ratio,
     )
-
-    sd = torch_model.state_dict()
-    reference_model.load_state_dict(sd)
-    reference_model.eval()
+    target_prefix = f"encoder.block.{block_i}.{efficient_self_attention_i}.attention."
+    reference_model = load_torch_model(
+        reference_model, target_prefix, module="semantic_sub", model_location_generator=model_location_generator
+    )
 
     torch_input_tensor = torch.reshape(torch_input_tensor, (batch_size, seq_len, hidden_size))
     torch_output = reference_model(torch_input_tensor, height, width)
