+++ models/demos/segformer/tests/pcc/test_segformer_overlap_path_embeddings.py
@@ -2,23 +2,16 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import torch
-import ttnn
-from ttnn.model_preprocessing import (
-    preprocess_model_parameters,
-    preprocess_layernorm_parameter,
-)
-from tests.ttnn.utils_for_testing import assert_with_pcc
-from transformers import SegformerModel
 import pytest
-from models.demos.segformer.tt.ttnn_segformer_overlap_patch_embeddings import (
-    TtSegformerOverlapPatchEmbeddings,
-)
+import torch
+from ttnn.model_preprocessing import preprocess_layernorm_parameter, preprocess_model_parameters
 
-from models.demos.segformer.reference.segformer_overlap_patch_embeddings import (
-    SegformerOverlapPatchEmbeddings,
-)
+import ttnn
+from models.demos.segformer.common import load_torch_model
+from models.demos.segformer.reference.segformer_overlap_patch_embeddings import SegformerOverlapPatchEmbeddings
+from models.demos.segformer.tt.ttnn_segformer_overlap_patch_embeddings import TtSegformerOverlapPatchEmbeddings
 from models.utility_functions import skip_for_grayskull
+from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
 def create_custom_preprocessor(device):
@@ -66,20 +59,18 @@ def test_segformer_overlap_patch_embeddings(
     width,
     patch_emb_i,
     device,
-    reset_seeds,
-    is_ci_env,
+    model_location_generator,
 ):
     torch_input_tensor = torch.randn(batch_size, num_channels, height, width)
 
-    torch_model = SegformerModel.from_pretrained("nvidia/segformer-b0-finetuned-ade-512-512")
-
-    torch_model = torch_model.encoder.patch_embeddings[patch_emb_i]
     reference_model = SegformerOverlapPatchEmbeddings(
         patch_size=patch_size, stride=stride, num_channels=num_channels, hidden_size=hidden_size
     )
-    sd = torch_model.state_dict()
-    reference_model.load_state_dict(sd)
-    reference_model.eval()
+    target_prefix = f"encoder.patch_embeddings.{patch_emb_i}"
+    reference_model = load_torch_model(
+        reference_model, target_prefix, module="semantic_sub", model_location_generator=model_location_generator
+    )
+
     torch_output = reference_model(torch_input_tensor)
 
     parameters = preprocess_model_parameters(
