+++ tests/nightly/single_card/segformer/test_segformer_for_image_classification.py
@@ -1,0 +1,1 @@
+models/demos/segformer/tests/pcc/test_segformer_for_image_classification.py
